+++ lang/en/stats_en.c
@@ -1,12 +1,12 @@
 #include "../../include/stats_en.h"
 #include <stdio.h>
 
-extern short unsigned int d_type1, f_type2, c_type3, e_type4, h_type5, b_type6, i_type7, g_type8, a_type9;
+extern int d_type1, f_type2, c_type3, e_type4, h_type5, b_type6, i_type7, g_type8, a_type9;
 
 /* Function that contains the basic structure/form of each question in the test */
-void stats_form(char stat1[120], char stat2[120], char resp1, char resp2, short unsigned int *stat1_type, short unsigned int *stat2_type) {
+void stats_form(char stat1[120], char stat2[120], char resp1, char resp2, int *stat1_type, int *stat2_type) {
 	char resp;
-	short unsigned int i;
+	int i;
 	putchar('\n');
 
 	puts(stat1);
